Assignment/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
pillow
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic[email]